habit-ai-server/habit-ai-server/requirements.txt
flask
python-dotenv
openai>=1.12.0
